+++ packages/builder/builder-rspack-provider/tests/plugins/default.test.ts
@@ -4,12 +4,16 @@ import { BuilderPlugin } from '@/types';
 
 describe('applyDefaultPlugins', () => {
   it('should apply default plugins correctly', async () => {
+    const { NODE_ENV } = process.env;
+    process.env.NODE_ENV = 'development';
     const builder = await createBuilder({});
 
     const {
       origin: { bundlerConfigs },
     } = await builder.inspectConfig();
     expect(bundlerConfigs[0]).toMatchSnapshot();
+
+    process.env.NODE_ENV = NODE_ENV;
   });
 
   it('should apply default plugins correctly when prod', async () => {
@@ -29,6 +33,9 @@ describe('applyDefaultPlugins', () => {
 
 describe('tools.rspack', () => {
   it('should match snapshot', async () => {
+    const { NODE_ENV } = process.env;
+    process.env.NODE_ENV = 'development';
+
     class TestPlugin {
       readonly name: string = 'TestPlugin';
 
@@ -58,6 +65,8 @@ describe('tools.rspack', () => {
     } = await builder.inspectConfig();
 
     expect(bundlerConfigs[0]).toMatchSnapshot();
+
+    process.env.NODE_ENV = NODE_ENV;
   });
 });