+++ packages/builder/builder-webpack-provider/src/core/startDevServer.ts
@@ -1,29 +1,13 @@
 import {
   debug,
-  logger,
-  DEFAULT_PORT,
-  StartDevServerResult,
+  startDevServer as baseStartDevServer,
   StartDevServerOptions,
+  getDevServerOptions,
 } from '@modern-js/builder-shared';
-import { merge } from '@modern-js/utils/lodash';
-import type { ModernDevServerOptions } from '@modern-js/server';
 import { createCompiler } from './createCompiler';
 import { initConfigs, InitConfigsOptions } from './initConfigs';
 import type { Compiler, MultiCompiler } from 'webpack';
-
-async function printDevServerURLs(urls: Array<{ url: string; type: string }>) {
-  const { chalk } = await import('@modern-js/utils');
-  let message = 'Dev server running at:\n\n';
-
-  message += urls
-    .map(
-      ({ type, url }) =>
-        `  ${`> ${type.padEnd(10)}`}${chalk.cyanBright(url)}\n`,
-    )
-    .join('');
-
-  logger.info(message);
-}
+import { getDevMiddleware } from './devMiddleware';
 
 export async function createDevServer(
   options: InitConfigsOptions,
@@ -32,7 +16,6 @@ export async function createDevServer(
   customCompiler?: Compiler | MultiCompiler,
 ) {
   const { Server } = await import('@modern-js/server');
-  const { applyOptionsChain } = await import('@modern-js/utils');
 
   let compiler: Compiler | MultiCompiler;
   if (customCompiler) {
@@ -48,53 +31,18 @@ export async function createDevServer(
   debug('create dev server');
 
   const builderConfig = options.context.config;
-  const devServerOptions = applyOptionsChain(
-    {
-      hot: builderConfig.dev?.hmr ?? true,
-      watch: true,
-      client: {
-        port: port.toString(),
-      },
-      port,
-      liveReload: builderConfig.dev?.hmr ?? true,
-      devMiddleware: {
-        writeToDisk: (file: string) => !file.includes('.hot-update.'),
-      },
-      https: builderConfig.dev?.https,
-      // merge devServerOptions from serverOptions
-      ...serverOptions.dev,
-    },
-    builderConfig.tools?.devServer,
-  );
-
-  const defaultConfig: ModernDevServerOptions['config'] = {
-    output: {
-      path: builderConfig.output?.distPath?.root,
-      assetPrefix: builderConfig.output?.assetPrefix,
-      distPath: builderConfig.output?.distPath,
-    },
-    source: {
-      alias: {},
-      define: {},
-      globalVars: {},
-    },
-    html: {},
-    tools: {
-      babel: {},
-    },
-    server: {},
-    runtime: {},
-    bff: {},
-  };
+  const { config, devConfig } = await getDevServerOptions({
+    builderConfig,
+    serverOptions,
+    port,
+  });
 
   const server = new Server({
     pwd: options.context.rootPath,
-    compiler,
+    devMiddleware: getDevMiddleware(compiler),
     ...serverOptions,
-    dev: devServerOptions,
-    config: serverOptions.config
-      ? merge({}, defaultConfig, serverOptions.config)
-      : defaultConfig,
+    dev: devConfig,
+    config,
   });
 
   debug('create dev server done');
@@ -104,60 +52,7 @@ export async function createDevServer(
 
 export async function startDevServer(
   options: InitConfigsOptions,
-  {
-    compiler,
-    printURLs = true,
-    strictPort = false,
-    serverOptions = {},
-  }: StartDevServerOptions = {},
+  startDevServerOptions: StartDevServerOptions = {},
 ) {
-  logger.info('Starting dev server...');
-
-  if (!process.env.NODE_ENV) {
-    process.env.NODE_ENV = 'development';
-  }
-
-  const { getPort } = await import('@modern-js/utils');
-  const builderConfig = options.context.config;
-
-  const port = await getPort(builderConfig.dev?.port || DEFAULT_PORT, {
-    strictPort,
-  });
-
-  options.context.devServer = {
-    hostname: 'localhost',
-    port,
-  };
-
-  const server = await createDevServer(options, port, serverOptions, compiler);
-
-  await options.context.hooks.onBeforeStartDevServerHook.call();
-
-  debug('listen dev server');
-  await server.init();
-
-  return new Promise<StartDevServerResult>(resolve => {
-    server.listen(port, async (err: Error) => {
-      if (err) {
-        throw err;
-      }
-
-      debug('listen dev server done');
-
-      const { getAddressUrls } = await import('@modern-js/utils');
-      const protocol = builderConfig.dev?.https ? 'https' : 'http';
-      const urls = getAddressUrls(protocol, port);
-
-      if (printURLs) {
-        await printDevServerURLs(urls);
-      }
-
-      await options.context.hooks.onAfterStartDevServerHook.call({ port });
-      resolve({
-        port,
-        urls: urls.map(item => item.url),
-        server,
-      });
-    });
-  });
+  return baseStartDevServer(options, createDevServer, startDevServerOptions);
 }