+++ packages/runtime/plugin-testing/src/base/config/resolver.ts
@@ -7,13 +7,9 @@ const resolver = enhanceResolve.create.sync({
 
 const shouldResolveByEnhance = (url: string) => /^@[^/]+\/[^/]+\/.*/.test(url);
 
-const internalResolve = function (request: string, options: any) {
+module.exports = function (request: string, options: any) {
   if (shouldResolveByEnhance(request)) {
     return resolver(options.basedir, request);
   }
   return options.defaultResolver(request, options);
 };
-
-module.exports = internalResolve;
-
-export default internalResolve;